feat: merge gmail_batch_archive and gmail_archive_by_query into gmail_archive
