+++ assistant/src/__tests__/session-skill-tools.test.ts
@@ -1665,7 +1665,6 @@ const GMAIL_TOOL_NAMES = [
   "gmail_mark_read",
   "gmail_draft",
   "gmail_archive",
-  "gmail_batch_archive",
   "gmail_label",
   "gmail_batch_label",
   "gmail_trash",
